pyproject.toml
[project]
name = "menu-ocr"
version = "0.2.0"
description = "Menu OCR: Extract structured menu data from restaurant menu images using deep learning"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "kaggle>=1.8.3",
    "numpy>=2.4.1",
    "opencv-python<=4.6.0.66",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pytesseract>=0.3.13",
    "rich>=14.3.1",
    "shapely>=2.1.2",
    "easyocr",
    "torch",
    "torchvision",
    "transformers>=4.36.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.1.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "tqdm>=4.66.0",
    "paddlepaddle>=2.5.0",
    "paddleocr>=2.7.0",
    "datasets>=2.16.0",
    "evaluate>=0.4.0",
    "accelerate>=0.25.0",
    "timm>=0.9.0",
    "sentencepiece>=0.1.99",
]

[tool.uv]
package = true

[project.scripts]
menu-ocr = "src.main:main"
